Contact Management System
This code defines a Contact struct to store the name and email of a contact. The ContactManager class manages the contacts using a vector. It provides methods to add a contact, delete a contact, and display all contacts. The main function acts as the user interface where the user can choose options to interact with the contact management system.

You can compile and run this code in a C++ compiler to test the Contact Management System. It will allow you to add contacts, delete contacts, and display all the existing contacts.